Split cpu mask types to a separate header
