+++ hpx/runtime/threads/topology.hpp
@@ -13,24 +13,13 @@
 #include <hpx/config.hpp>
 #include <hpx/exception_fwd.hpp>
 #include <hpx/runtime/naming_fwd.hpp>
-#include <hpx/util/assert.hpp>
+#include <hpx/runtime/threads/cpu_mask.hpp>
 
-#include <boost/cstdint.hpp>
 #include <boost/thread/thread.hpp>
 
 #include <cstddef>
 #include <iosfwd>
-#include <limits>
 #include <utility>
-
-#if defined(HPX_HAVE_MORE_THAN_64_THREADS) || (defined(HPX_HAVE_MAX_CPU_COUNT) \
-            && HPX_HAVE_MAX_CPU_COUNT > 64)
-# if defined(HPX_HAVE_MAX_CPU_COUNT)
-#include <bitset>
-# else
-#include <boost/dynamic_bitset.hpp>
-# endif
-#endif
 
 #if defined(HPX_HAVE_HWLOC)
 #include <hpx/error_code.hpp>
@@ -41,190 +30,6 @@
 
 namespace hpx { namespace threads
 {
-    /// \cond NOINTERNAL
-#if !defined(HPX_HAVE_MORE_THAN_64_THREADS) || (defined(HPX_HAVE_MAX_CPU_COUNT) \
-             && HPX_HAVE_MAX_CPU_COUNT <= 64)
-    typedef boost::uint64_t mask_type;
-    typedef boost::uint64_t mask_cref_type;
-
-    inline boost::uint64_t bits(std::size_t idx)
-    {
-       HPX_ASSERT(idx < CHAR_BIT * sizeof(mask_type));
-       return boost::uint64_t(1) << idx;
-    }
-
-    inline bool any(mask_cref_type mask)
-    {
-        return mask != 0;
-    }
-
-    inline mask_type not_(mask_cref_type mask)
-    {
-        return ~mask;
-    }
-
-    inline bool test(mask_cref_type mask, std::size_t idx)
-    {
-        HPX_ASSERT(idx < CHAR_BIT * sizeof(mask_type));
-        return (bits(idx) & mask) != 0;
-    }
-
-    inline void set(mask_type& mask, std::size_t idx)
-    {
-        HPX_ASSERT(idx < CHAR_BIT * sizeof(mask_type));
-        mask |= bits(idx);
-    }
-
-    inline std::size_t mask_size(mask_cref_type mask)
-    {
-        return CHAR_BIT * sizeof(mask_type);
-    }
-
-    inline void resize(mask_type& mask, std::size_t s)
-    {
-        HPX_ASSERT(s <= CHAR_BIT * sizeof(mask_type));
-    }
-
-    inline std::size_t find_first(mask_cref_type mask)
-    {
-        if (mask) {
-            std::size_t c = 0;    // Will count mask's trailing zero bits.
-
-            // Set mask's trailing 0s to 1s and zero rest.
-            mask = (mask ^ (mask - 1)) >> 1;
-            for (/**/; mask; ++c)
-                mask >>= 1;
-
-            return c;
-        }
-        return ~std::size_t(0);
-    }
-
-    inline bool equal(mask_cref_type lhs, mask_cref_type rhs, std::size_t)
-    {
-        return lhs == rhs;
-    }
-
-    // return true if at least one of the masks has a bit set
-    inline bool bit_or(mask_cref_type lhs, mask_cref_type rhs, std::size_t)
-    {
-        return (lhs | rhs) != 0;
-    }
-
-    // return true if at least one bit is set in both masks
-    inline bool bit_and(mask_cref_type lhs, mask_cref_type rhs, std::size_t)
-    {
-        return (lhs & rhs) != 0;
-    }
-
-#define HPX_CPU_MASK_PREFIX "0x"
-
-#else
-# if defined(HPX_HAVE_MAX_CPU_COUNT)
-    typedef std::bitset<HPX_HAVE_MAX_CPU_COUNT> mask_type;
-    typedef std::bitset<HPX_HAVE_MAX_CPU_COUNT> const& mask_cref_type;
-# else
-    typedef boost::dynamic_bitset<boost::uint64_t> mask_type;
-    typedef boost::dynamic_bitset<boost::uint64_t> const& mask_cref_type;
-# endif
-
-    inline bool any(mask_cref_type mask)
-    {
-        return mask.any();
-    }
-
-    inline mask_type not_(mask_cref_type mask)
-    {
-        return ~mask;
-    }
-
-    inline bool test(mask_cref_type mask, std::size_t idx)
-    {
-        return mask.test(idx);
-    }
-
-    inline void set(mask_type& mask, std::size_t idx)
-    {
-        mask.set(idx);
-    }
-
-    inline std::size_t mask_size(mask_cref_type mask)
-    {
-        return mask.size();
-    }
-
-    inline void resize(mask_type& mask, std::size_t s)
-    {
-# if defined(HPX_HAVE_MAX_CPU_COUNT)
-        HPX_ASSERT(s <= mask.size());
-# else
-        return mask.resize(s);
-# endif
-    }
-
-    inline std::size_t find_first(mask_cref_type mask)
-    {
-# if defined(HPX_HAVE_MAX_CPU_COUNT)
-        if (mask.any())
-        {
-            for (std::size_t i = 0; i != HPX_HAVE_MAX_CPU_COUNT; ++i)
-            {
-                if (mask[i])
-                    return i;
-            }
-        }
-        return ~std::size_t(0);
-# else
-        return mask.find_first();
-# endif
-    }
-
-# if defined(HPX_HAVE_MAX_CPU_COUNT)
-#define HPX_CPU_MASK_PREFIX "0b"
-#else
-#define HPX_CPU_MASK_PREFIX "0x"
-#endif
-
-    inline bool equal(mask_cref_type lhs, mask_cref_type rhs, std::size_t numbits)
-    {
-        for (std::size_t j = 0; j != numbits; ++j)
-        {
-            if (test(lhs, j) != test(rhs, j))
-            {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    // return true if at least one of the masks has a bit set
-    inline bool bit_or(mask_cref_type lhs, mask_cref_type rhs, std::size_t numbits)
-    {
-        for (std::size_t j = 0; j != numbits; ++j)
-        {
-            if (test(lhs, j) || test(rhs, j))
-            {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    // return true if at least one bit is set in both masks
-    inline bool bit_and(mask_cref_type lhs, mask_cref_type rhs, std::size_t numbits)
-    {
-        for (std::size_t j = 0; j != numbits; ++j)
-        {
-            if (test(lhs, j) && test(rhs, j))
-            {
-                return true;
-            }
-        }
-        return false;
-    }
-#endif
-    /// \endcond
-
     struct HPX_EXPORT topology
     {
         virtual ~topology() {}
